requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
langchain>=0.2.16
langchain-core>=0.2.38
langchain-openai>=0.1.22
langchain-text-splitters>=0.2.4
langchain-chroma>=0.1.3
langchain-huggingface>=0.0.3
langgraph>=0.2.14
chromadb>=0.5.5
sentence-transformers>=3.0.1
torch>=2.4.0
pandas>=2.2.2
requests>=2.32.3
beautifulsoup4>=4.12.3
typing-extensions>=4.12.2
